First Pins Implementation
This adds the ability to render pins and move them around, along with
a button to add them (which will of course be expanded on).  At the
moment they respect boundaries, both of the board edge and each other,
and they snap to guides at both center, left and right edges.