Style Changes
Just a few final style changes before starting in on the backend.  I
changed the header to better fit the rest of the design (if you can
call it a design), changed the favicon, cleaned up the manifest, and
changed the color of the center guideline (if there is one) to black.