Moved globalContext to Store
When I started this project, I thought I'd keep the state management
simple by just putting each part of it into its own little context
provider.  This has turned out to way over-complicate things, as a
bunch of parts end up needing to pull in multiple contexts anyways and
it would be ideal to be able to update multiple parts of the state
simultaneously which is a mess right now with multiple separate states
being maintained.

So now I'm moving everything to a centralized store with useReducer.
This is the first part of that migration, which just sets up the basic
scaffolding for it and replaces globalContext with the store.  The
following commits will move all the other contexts over to the store
one at a time until everything is centralized.