Add Straight Cutter Diameter
I forgot an important detail when initially setting up the UI, which
is that the diameter of the straight cutter also matters.  We have to
have a gap between the bases of dovetails of at least the straight
cutter diameter, since the cutter will have to get into that space to
clear out the waste around the pins.  This adds a setting for the
straight cutter and updates all the settings modules to properly
account for it when setting minimums/maximums.