Rendered Pin Board A Side
This renders the A version of the pins board.  I also tested the files
on the origin today, and discovered that the previous style didn't
work.  So instead I've created inline style strings for pockets and
guides and I'm using those on the paths.  Together with an XML header,
the origin now seems to be happy to load the files.