Pin Creator Bugfix (Again)
So fixing the previous bug in the pin creator introduced a bigger one:
if you don't have enough space left for whatever size you last
created, it disables the text box so you can't adjust to a smaller
size.  This fixes it so it only grays out the input when you can't fit
any new pins in.