(Kind Of) Add B Side Pin Rendering
This adds rendering of the B side of the pin board, which I thought
was working...but then after actually cutting some test boards I
realized that I got the trigonometry wrong and undersized the pins
significantly.  So I'm committing this for now, but I need to figure
out where I went wrong in my calculation of the max/min widths for the
pins with margin space.