Fix Trig Again
Okay, so turns out the last fix only fixed the A side of the pins
board.  Turns out that when your pin isn't centered between the middle
and top lines, flipping the X positions on those lines does not give
you the same sized pins.  Whoops.  This latest change takes into
account the orientation of the pin when calculating the expanded
max/min widths, and I've tested it and actually gotten tight joints
all around.