Minimum Pin Width Updates
Two changes here.  First, I increased the minimum size of a pin to
0.1mm + the diameter of the dovetail bit.  That's because I found that
the origin refuses to allow me to fit the 1/2in dovetail bit into the
fingers if they're exactly 1/2in wide.  I haven't done the same with
the pins, because even at minimum distance you can still get the cut
made with the parts that you are allowed to cut.