Added State Validation
This adds some validation functions that run on state changes.  They
remove any pins that, due to global parameters changing, have become
illegal.  Pins that have gone out of bounds, half pins that are now
too long for the material, or pins that are too close to each other
for the straight cutter diameter or too narrow for the dovetail
diameter.