Added Custom Anchor Support
This adds a customizable red right triangle at one of the material
corners.  I don't have the newest firmware update for the Origin so I
can't test this yet, but it _should_ create a custom anchor at the
given point so you can just anchor directly to the origin point of
your grid.