Added Auto-Layout Tools
This adds a tool to automatically distribute pins across the available
space, not including the space already occupied by half pins.  It has
three modes, with fixed pin and tail widths being fairly
straightforward.  The even spacing mode was kind of a crash review of
high school algebra, but I think I got all the calculations right if
not necessarily elegant.

The biggest problem here is that I didn't get the min/max counts and
sizes all perfectly correct in the UI.  The math to make them precise
would have been even hairier, so I built in a little bit of buffer by
using simpler math in a few places, but it should be good enough for
practical purposes.

Of course the instructions are now out of sync, but I'll update that
in the next commit before deploying and then starting work on
half-blind dovetails.