Limit Updates for Half-Blind Dovetails
This diff begins laying the foundation for half-blind dovetails.  Now
we have an extra selector in the general settings section that lets us
switch between through and half-blind dovetails, and it toggles
material thickness to dovetail depth instead. Then I had to add a new
input for the actual thickness of the physical material which will be
used in the template generation.  This is going to be a little
confusing in the split between UI and code, and I _should_ probably go
back and rename the existing value, but it's just not worth it
considering how little work is left to get this thing into its final
state.  Or maybe not, in which case whoever needs to modify it next
will at least have this commit message to look back on and be mildly
annoyed at me for.

The first thing I discovered after updating the useLimits helper is
that the minX/maxX values in the visualizer don't currently use it, so
they broke spectacularly right away.  I fixed that though, so now I
_believe_ the pin layout UI is consistent for half-blind (which
requires that the pins not be spaced closer together than the dovetail
cutter allows, since they'll now be cut with the dovetail cutter).