Half-Blind Export UI
This adds a UI for exporting half-blind dovetails, which includes A
and B variants of the pin boards and an input for glue gap, which will
need to be baked into the templates for half blinds due to the fact
that adding margin to the cut will also screw up the depth on the
tails templates.  While I was at it, I removed the user-configurable
buffer from both versions and replaced it with one that's
automatically calculated and has no guide.  This will simplify the
UI/templates and the buffer is no longer needed now that we have
custom anchors.