Tail Generation for Half Blind Mode
This adds generation of A and B side tails templates for half-blind
dovetails.  I haven't actually tested any of the generated templates
yet, so there may need to be some more changes, but as far as I can
tell the math seems right.  I realized along the way that I actually
need to have the full dovetail bit spacing in between pins, because
otherwise there won't be enough room for the rounded cut when cutting
out the tails.