Fixed Tail Radius
I originally made the curves on the tail boards the same radius as the
cutter, but this was a mistake.  The part of the tail that's going to
interface with the pins board is the top, which is cut farther up the
cutter.  So if I cut to the cutter's radius at the bottom, I get a
larger radius at the top.  This adjusts the path to account for that
difference, and removes the change I had made to increase the minimum
pin spacing.