Half-blind Pins Generation
This adds generation of pins boards in half blind mode.  I also added
an extra depth parameter to enable the user to add some depth to the
holes cut in the pins board, which will leave the ends a little bit
proud so they can be planed or sanded flush.  This completes the
half-blind mode, I'll just need to update the instructions before
publishing.