Fix Unit Conversion in Numeric Inputs
Previously I was passing min/max values in mm to the HTML <input>
elements, even when the tool is in inch mode.  This causes problems
when users try to enter values that would be okay in inches but out of
range in mm.  This just applies a conversion factor to those
attributes when in imperial mode.