Updated Instructions
This updates the instructions to account for the latest changes
including the half-blind mode.  At this point I'm considering the
project feature complete and don't expect to make any further changes
aside from bug fixes.  Although of course I also said that when I
finished the original version with through dovetails, so...we'll see.