TS Transition: General Store
I want to make some more big changes to the app but it's already
getting pretty unwieldy to work with, so I'm doing what I should have
done in the first place and migrating to typescript.  Starting with
the context code, which is probably going to be the most complex to
migrate.  This moves over the General sub-store, and adds a few hacky
fixes to store.js to make it work with the new style of action which
is going to have separate `store` and `type` fields to make it easier
to route them to the correct resolver in a type-safe way.

I'll do a few more commits to get the individual stores moved over,
and then one probably trickier one to get the store itself ported.