TS Transition: Guides
This is a relatively simple conversion of the Guides context store to
typescript.  With only a few fields, no nested values, and a single
action, this one was basically trivial.

I also realized along the way that the actions in the General store
should have used partials instead of expecting the complete type for
updates, so I changed that while I was at it.