TS Transition: Pins, HalfPins, Pin Utils
This one ended up slightly more involved than I'd planned for, it was
a bit of a tug on the string and more than expected falls apart
situation.  I was just doing the pins store, but that had some
dependencies on the half pins store and the pin utils file that
required converting them both to TS to get everything to build nicely.
So now all that's left in context is to convert over the store
provider itself.