TS Transition: Completed the Store
This completes the migration of everything in src/context to
typescript.  All the store types are now typed.

I pulled in ts-deepmerge to do the merge from URI state in a type-safe
way, but unfortunately the npm package didn't include source which
screwed up the source map build.  To work around that, since it's a
one-file package I just added the file to this repo, with a note at
the top about licensing since the original author didn't include
copyright headers and only specified the license on npm.