Fix Edge Spacing
When constructing the cut areas the generator previously defaulted to
adding 1.1 times the cutter radius to the left and right side of the
material.  This is a problem when cutting tails with half pins or pins
*without pins* if the cutter needs to come especially close to the
edge of the material, as there may not be enough to accommodate its
entire diameter.

What I probably should have done here is just switched to using 1.1 *
the diameter instead of the radius, but I decided to be clever about
it and figure out the actual amount of clearance needed if it's
greater than the existing amount and use that instead.  I'm sure this
will come back to bite me at some point in the future, at which point
I'll revert this commit, switch to the simpler solution, and tell
myself I told you so, but for now let's see how this works.