Restore Material Thickness on Switch Back to Through Mode
When switching from through to half-blind mode we do a little trick
where the material thickness gets transferred to dovetailLength and
thickness gets set to 2/3 of the existing thickness, since "thickness"
in half-blind mode actually means the cut depth.  This resulted in
ever-decreasing material thickness if the user switched back and forth
between the two modes, so this just restores the original material
thickness from the dovetailLength field if that happens.