Pin Sorting Fix
The pin and tail generation code for through dovetails didn't sort the
pins prior to iterating through them, which created the possibility of
out of order pin rendering.  Oddly enough this hasn't really been a
problem before, I think it may depend on the order pins are created
and deleted.